requirements.txt
pyspark~=3.5.3
colorlog~=6.9.0
dune_client~=1.7.7
python-dotenv~=1.0.1
requests~=2.32.3
pandas~=2.2.3
behave~=1.2.6